Refactoring music panel
